+++ default.nix
@@ -22,6 +22,8 @@ in {
   inherit (haskellPackages)
     hnix-store-core
     hnix-store-db
+    hnix-store-nar
+    hnix-store-readonly
     hnix-store-remote
     hnix-store-tests;
   haskellPackages = lib.dontRecurseIntoAttrs haskellPackages;