+++ hnix-store-core/src/System/Nix/Store/Types.hs
@@ -1,0 +1,24 @@
+module System.Nix.Store.Types
+  ( FileIngestionMethod(..)
+  , PathFilter(..)
+  , RepairMode(..)
+  ) where
+
+import GHC.Generics (Generic)
+
+-- | Add path recursively or not
+data FileIngestionMethod
+  = FileIngestionMethod_Flat
+  | FileIngestionMethod_FileRecursive
+  deriving (Bounded, Eq, Generic, Enum, Ord, Show)
+
+-- | Path filtering function
+newtype PathFilter = PathFilter
+  { pathFilterFunction :: FilePath -> Bool
+  }
+
+-- | Repair mode
+data RepairMode
+  = RepairMode_DoRepair
+  | RepairMode_DontRepair
+  deriving (Bounded, Eq, Generic, Enum, Ord, Show)