+++ hnix-store-readonly/src/System/Nix/Store/ReadOnly.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module System.Nix.ReadonlyStore
+module System.Nix.Store.ReadOnly
   ( makeStorePath
   , makeTextPath
   , makeFixedOutputPath
@@ -13,7 +13,8 @@ import Crypto.Hash (Context, Digest, SHA256)
 import Data.ByteString (ByteString)
 import Data.HashSet (HashSet)
 import System.Nix.Hash (BaseEncoding(Base16), NamedAlgo(algoName))
-import System.Nix.StorePath (StoreDir, StorePath(StorePath), StorePathName)
+import System.Nix.Store.Types (FileIngestionMethod(..), PathFilter, RepairMode)
+import System.Nix.StorePath (StoreDir, StorePath, StorePathName)
 
 import qualified Crypto.Hash
 import qualified Data.ByteString.Char8
@@ -34,7 +35,8 @@ makeStorePath
   -> Digest hashAlgo
   -> StorePathName
   -> StorePath
-makeStorePath storeDir ty h nm = StorePath storeHash nm
+makeStorePath storeDir ty h nm =
+ System.Nix.StorePath.unsafeMakeStorePath storeHash nm
  where
   storeHash = System.Nix.StorePath.mkStorePathHashPart @hashAlgo s
   s =
@@ -66,20 +68,21 @@ makeFixedOutputPath
   :: forall hashAlgo
   .  NamedAlgo hashAlgo
   => StoreDir
-  -> Bool
+  -> FileIngestionMethod
   -> Digest hashAlgo
   -> StorePathName
   -> StorePath
 makeFixedOutputPath storeDir recursive h =
-  if recursive && (algoName @hashAlgo) == "sha256"
-    then makeStorePath storeDir "source" h
-    else makeStorePath storeDir "output:out" h'
+  if recursive == FileIngestionMethod_FileRecursive
+     && (algoName @hashAlgo) == "sha256"
+  then makeStorePath storeDir "source" h
+  else makeStorePath storeDir "output:out" h'
  where
   h' =
     Crypto.Hash.hash @ByteString @SHA256
       $  "fixed:out:"
       <> Data.Text.Encoding.encodeUtf8 (algoName @hashAlgo)
-      <> (if recursive then ":r:" else ":")
+      <> (if recursive == FileIngestionMethod_FileRecursive then ":r:" else ":")
       <> Data.Text.Encoding.encodeUtf8 (System.Nix.Hash.encodeDigestWith Base16 h)
       <> ":"
 
@@ -96,12 +99,15 @@ computeStorePathForPath
   :: StoreDir
   -> StorePathName        -- ^ Name part of the newly created `StorePath`
   -> FilePath             -- ^ Local `FilePath` to add
-  -> Bool                 -- ^ Add target directory recursively
-  -> (FilePath -> Bool)   -- ^ Path filter function
-  -> Bool                 -- ^ Only used by local store backend
+  -> FileIngestionMethod  -- ^ Add target directory recursively
+  -> PathFilter           -- ^ Path filter function
+  -> RepairMode           -- ^ Only used by local store backend
   -> IO StorePath
 computeStorePathForPath storeDir name pth recursive _pathFilter _repair = do
-  selectedHash <- if recursive then recursiveContentHash else flatContentHash
+  selectedHash <-
+    if recursive == FileIngestionMethod_FileRecursive
+      then recursiveContentHash
+      else flatContentHash
   pure $ makeFixedOutputPath storeDir recursive selectedHash name
  where
   recursiveContentHash :: IO (Digest SHA256)