+++ hnix-store-remote/src/System/Nix/Store/Remote.hs
@@ -27,6 +27,7 @@ module System.Nix.Store.Remote
   , runStore
   , syncWithGC
   , verifyStore
+  , module System.Nix.Store.Types
   , module System.Nix.Store.Remote.Types
   ) where
 
@@ -43,6 +44,7 @@ import qualified System.Nix.Hash
 import qualified Data.ByteString.Lazy          as BSL
 
 import System.Nix.Derivation (Derivation)
+import System.Nix.Store.Types (FileIngestionMethod(..), RepairMode(..))
 import           System.Nix.Build               ( BuildMode
                                                 , BuildResult
                                                 )
@@ -80,18 +82,21 @@ addToStore
    . (NamedAlgo a)
   => StorePathName        -- ^ Name part of the newly created `StorePath`
   -> NarSource MonadStore -- ^ provide nar stream
-  -> Recursive            -- ^ Add target directory recursively
-  -> RepairFlag           -- ^ Only used by local store backend
+  -> FileIngestionMethod  -- ^ Add target directory recursively
+  -> RepairMode           -- ^ Only used by local store backend
   -> MonadStore StorePath
 addToStore name source recursive repair = do
-  Control.Monad.when (unRepairFlag repair)
+  Control.Monad.when (repair == RepairMode_DoRepair)
     $ error "repairing is not supported when building through the Nix daemon"
 
   runOpArgsIO AddToStore $ \yield -> do
     yield $ BSL.toStrict $ Data.Binary.Put.runPut $ do
       putText $ System.Nix.StorePath.unStorePathName name
-      putBool $ not $ System.Nix.Hash.algoName @a == "sha256" && (unRecursive recursive)
-      putBool (unRecursive recursive)
+      putBool
+        $ not
+        $ System.Nix.Hash.algoName @a == "sha256"
+          && recursive == FileIngestionMethod_FileRecursive
+      putBool (recursive == FileIngestionMethod_FileRecursive)
       putText $ System.Nix.Hash.algoName @a
     source yield
   sockGetPath
@@ -104,10 +109,11 @@ addTextToStore
   :: Text              -- ^ Name of the text
   -> Text              -- ^ Actual text to add
   -> HashSet StorePath -- ^ Set of `StorePath`s that the added text references
-  -> RepairFlag        -- ^ Repair flag, must be `False` in case of remote backend
+  -> RepairMode        -- ^ Repair mode, must be `RepairMode_DontRepair` for remote backend
+                       --   (only valid for local store)
   -> MonadStore StorePath
 addTextToStore name text references' repair = do
-  Control.Monad.when (unRepairFlag repair)
+  Control.Monad.when (repair == RepairMode_DoRepair)
     $ error "repairing is not supported when building through the Nix daemon"
 
   storeDir <- getStoreDir
@@ -325,7 +331,7 @@ syncWithGC :: MonadStore ()
 syncWithGC = Control.Monad.void $ simpleOp SyncWithGC
 
 -- returns True on errors
-verifyStore :: CheckFlag -> RepairFlag -> MonadStore Bool
+verifyStore :: CheckFlag -> RepairMode -> MonadStore Bool
 verifyStore check repair = simpleOpArgs VerifyStore $ do
   putBool $ unCheckFlag check
-  putBool $ unRepairFlag repair
+  putBool $ repair == RepairMode_DoRepair